LayoutTests/http/tests/security/anchor-download-block-crossorigin-expected.txt

CONSOLE MESSAGE: line 133: The download attribute on anchor was ignored because its href URL has a different security origin.
Tests that the download attribute is ignored if the link is cross origin.

It should navigate the subframe instead of downloading the file.
